// tradebyte_frontend/generated/models/AdDtoPagedResult.ts
/**
 * TradeByte API
 * Közösségi adok-veszek oldal API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AdDto } from '../models/AdDto';
import { HttpFile } from '../http/http';

export class AdDtoPagedResult {
    'items'?: Array<AdDto> | null;
    'total'?: number;
    'page'?: number;
    'pageSize'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<AdDto>",
            "format": ""
        },
        {
            "name": "total",
            "baseName": "total",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "page",
            "baseName": "page",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "pageSize",
            "baseName": "pageSize",
            "type": "number",
            "format": "int32"
        }    ];

    static getAttributeTypeMap() {
        return AdDtoPagedResult.attributeTypeMap;
    }

    public constructor() {
    }
}